cloud-functions/beans/requirements.txt
functions-framework==3.*
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-storage
kfp
